+++ tests/test-post-index-helpers.php
@@ -297,4 +297,16 @@ class Post_Index_Helpers_Test extends WP_UnitTestCase {
 		}
 	}
 
+	/* c2c_the_index() */
+
+	public function test_the_index_for_valid_index() {
+		$this->loop( 3 );
+
+		$this->expectOutputRegex( '/^2$/', c2c_the_index() );
+	}
+
+	public function test_the_index_for_invalid_index() {
+		$this->expectOutputRegex( '/^-1$/', c2c_the_index() );
+	}
+
 }