+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS
+* Change: Unit tests: Add tests for `c2c_the_index()`
 
 ## 1.2.9 _(2019-12-23)_
 * Change: Note compatibility through WP 5.3+